Layers work now! No more errors. Made sure that the display works properly too.
Need to fix existing tests and add new tests for layers.
Also need to add layer functionality to the EditorView itself - list of layers that controls a list of shapes?

Also need to do some groundwork for rotation! In particular, add a field to Motion.
(Do this through a new constructor to avoid breaking existing code?)